core/pollfds: Use direct indexing instead of searches
Convert pollfds to act more like select, where a given fd
can be used directly as an index into the fdset.  This avoids
the linear searches currently done in the code to find the
fd in the fdset array.

Entries that are invalid are marked with an fd of -1, which
signals that poll() should ignore the entry.  This change
slightly increases the size of the fdset allocation, but
assumes that most entries will be used to reference a valid
fd value.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>