core/pollfds: Always process work_item_list
In ofi_pollfds_wait, we call poll() prior to processing the
work item list.  If wait is called with a timeout of 0, the
work item list may never be processed as poll may always
return 0.

This impacts the tcp provider which uses the pollfds.  The
tcp provider toggles a socket's POLLOUT state based on if
there's pending data to send.  However, the update may
never be processed, resulting in the socket never being
checked if it's now safe to send data.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>