prov/efa: rename and add comment to rxr_pkt_rtm_init_rx_entry()
This patch rename rxr_pkt_rtm_init_rx_entry() to
rxr_pkt_update_rx_entry() because the new name fits its behavior
better: it is not only applied on a newly allocated unexpected
rx_entry, but also applied to existing rx_entry when it is matched
to an RTM packet.

This patch also added comments for this function.

Signed-off-by: Wei Zhang <wzam@amazon.com>