prov/cxi: Ack events when exiting CQ progress
When the EQ size is small and the ACK batch size is large compared to
the EQ size, future events can be dropped in this configuration due to
EQ space potentially not being ACK'ed due to batch size not being met.
To prevent this issue, allows ACK the EQ when exiting progress to unsure
space is available.

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>