prov/tcp: Replace multiple buffer pools with one
TCP maintains an arary of 5 buffer pools, but the pools
contain identical structures, with only 1 8-bit value
differing between the structures.  Collapse these into
1 pool.

This removes the use of op_data fields (which is carried
in the base_hdr) to indicate which buffer pool an entry
came from.  We maintain the only needed op_data field
actually used by the protocol for wire compatibility.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>