prov/tcp: Define common xfer entry allocation path for user transmits
We'll need to limit the number of outstanding transmits that a
user can post to the endpoint.  Create common xfer_entry_alloc
like paths, so that we can isolate the tracking to a single
location.

This restructures and updates the xfer_entry allocation and free
routines, moving those to static inline calls.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>