prov/efa: simplify zero copy receive workflow
Current zero copy receive workflow:

When application call fi_recv,
  1. create a rx_entry, put it in ep->rx_list,
  2. construct a pkt_entry using user buffer, post it to device.

When we got receive completion
  1. find the rx_entry in ep->rx_list
  2. use rx_entry to write rx completion.

The process of put rx_entry in ep->rx_list, and later get it
back is unnecessary. Because pkt_entry and rx_entry was created
at the same time, we can set pkt_entry->x_entry to rx_entry when
it is created.

This patch simplify the process to:

When application call fi_recv,
  1. create a rx_entry.
  2. construct a pkt_entry using user buffer.
  3. set pkt_entry->x_entry to rx_entry

When we get receive completion:
  1. get rx_entry from pkt_entry->x_entry
  2. use the rx_entry to write rx completion.

Signed-off-by: Wei Zhang <wzam@amazon.com>