prov/hook: Add new hooking provider dmabuf_peer_mem
The kernel module dmabuf_peer_mem.ko is a IB peer-memory client for dmabuf. It
maintains a registry of dmabuf entries, each of which containing information
such as base address, size, and dmabuf fd. With this registry, it can respond
to requests from the MOFED RDMA driver to perform virtual address to DMA
address mapping. The registry is updated by the user space applications via
the character device /dev/dmabuf_reg.

This new hooking provider intercepts memory registration calls. It checks if
the buffer to be registered is backed by dmabuf and if so, saves the info to
the registry maintained by the dmabuf_peer_mem.ko kernel module.

This hooking provider only applies to core providers to avoid duplicated action.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>