prov/tcp: Remove byte-swapping of send header after transfers
Once a transmit has been marked done, we byte swap the header
back to host order, so that completion handling can read the
data.  However, when async sends are added, we can't modify
the header until the async operation completes.  There's no
actual need to access the send header after the transfer has
completed, so remove the byte swap code.  Update the completion
reporting code to only use the header on received messages.
For send completions, the fields in the cqe being filled out
are not defined, to set to 0.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>