fabtests: add functional/rdm_multi_client test.
This program tests the functionality of RDM endpoint in the case
that a persistent server does ping-pong with multiple clients that
come and leave in sequence. The client connects to a server, sends
ping-pong, disconnects with the server by cleaning all fabric
resources, and repeats.
If the `-R` option is specified, it will re-use the first client's
address for the subsequent clients by setting the src_addr for
endpoints 2..n to the output of fi_getname() of the first client.

Signed-off-by: Shi Jin <sjina@amazon.com>