[v1.12.x] prov/efa: do not use delivery complete for inject
If application called fi_inject, it will not receive a completion
entry for the message, which means we do not need to enable
delivery complete for this message either.

This patch makes the change that delviery complete is disabled
for inject, which helps improve performance.

Signed-off-by: Wei Zhang <wzam@amazon.com>

Cherry picked from 92b3dc1cb