prov/efa: identify and handle QP reuse
QP reuse means a peer has the same gid+qpn but different qkey was
inserted to address vector. It means the old QP has been destroyed
and new one was created with same qpn.

To identify such case, when inserting an address to address vector,
we first find the AH from gid using av->ah_map. Then search the
reverse_av hash table, which uses ahn+qpn as key. If an existing
entry is found in reverse_av, QP reuse is detected. In this case,
we save the old qkey in new peer for future use, then call
efa_conn_release() to release the resource of the old peer.

Signed-off-by: Wei Zhang <wzam@amazon.com>