configure, hmem_ze: separate ZE and DRM dependencies
Some of the ZE code requires the i915 DRM headers. The code
path that needs this is the IPC code. Not all ZE use cases
need this support and even the shm provider has a fallback
if it is not available (SAR).
Separate these dependencies to allow ZE to build with or
without this header.

Signed-off-by: Alexia Ingerson <alexia.ingerson@intel.com>