core: Maintain list of all providers
If multiple versions of a provider are present, include it in the
list of available provider.  Do not discard the provider just
because it has a lower version number.  This allows loading
different versions of the same provider in case the application
requires a specific version (for testing or features).  To
maintain prior behavior of only exposing the latest version of a
provider, older versions are marked as hidden.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>