fabtests/runfabtest.sh: skip ubertest if not config file found
If -u is not provided, runfabtests will attempt to find a suitable config
file. If it is not able to find a config file, runfabtests will still run
by passing -u {NULL} into the test causing a hang. Check for a null config
file and return after setting the config file.

Signed-off-by: aingerson <alexia.ingerson@intel.com>