prov/rxm: Fix calling fi_getinfo with FI_SOURCE but null params
If the application specifies a null node but a service, rxm
marks this as a wildcard listen.  It then sets service to null.
This results in fi_getinfo failing (FI_SOURCE set, but node and
service are null).  The problem is reproducible using fi_av_xfer
and only specifying the destination address.

In this case, we need to clear the FI_SOURCE flag, so that the
msg provider can return a list of usable addresses.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>