prov/shm: fix IPC iface used for IPC
When going through the IPC protocol, shm should be using the interface from the
command/src, not the destination buffer which could be host memory and will not
point to the correct iface calls.

Assert and make a note to fix incompatible interfaces for IPC. We cannot copy from
one type of device to another. They either have to be the same or the destination
has to be host memory.

Signed-off-by: aingerson <alexia.ingerson@intel.com>