prov/efa: add connid to CTS packet header.
This patch introduce a new usage to 4 byte space
in rxr_cts_hdr, which is currently named "padding"

If the RXR_PKT_CONNID_HDR flag is toggled in "flags",
the space is used for "connid" (sender's connection ID)

The patch then "connid" in rxr_init_cts_hdr(), also toggle
on the RXR_PKT_CONNID_HDR flag in "flags".

A section is added in rxr_pkt_connid_ptr(), which return the
ptr to connid in CTS packts if RXR_PKT_CONNID_HDR is on.

The macro RXR_CTS_HDR_SIZE was removed for simplicity

Signed-off-by: Wei Zhang <wzam@amazon.com>