prov/efa: add connid to RECEIPT packet
This patch introduce a new usage to the 4 bytes
padding space in rxr_receipt_hdr, which is currently
named "padding"

If the RXR_PKT_CONNID_HDR is toggled in "flags", the
space is used to store "connid" (sender's connection ID)

The patch sets "connid" and toggle on the RXR_PKT_CONNID_HDR
flag in "flags".

It added a section in rxr_pkt_connid_ptr() to return connid ptr
in RECEIPT packet if RXR_PKT_CONNID_HDR is on.

Signed-off-by: Wei Zhang <wzam@amazon.com>