prov/efa: add connid to DATA packet if peer needs it
This patch added optional connid header to data packet if
the peer needs it.

A struct named rxr_data_opt_connid_hdr was defined as
an optional connid header of DATA packet.

A flag RXR_DATA_OPT_CONNID_HDR was introduced to indicate
the presence of the connid hdr in a DATA packet.

A section was added to rxr_pkt_connid_ptr() to return pointer
to connid in header for DATA packet if RXR_DATA_OPT_CONNID_HDR
is on.

This patch removed the type "struct rxr_data_pkt" because
DATA packet's header is not fixed anymore.

This patch also did a refactor:

it removed the function rxr_pkt_post_data(), and introduced
function rxr_pkt_init_data(). With this change, the patch
used rxr_pkt_post_ctrl() to send DATA packet.

Signed-off-by: Wei Zhang <wzam@amazon.com>