prov/efa: add connid to READRSP packet header
This patch introduced a new usage of the 4 bytes space in
rxr_readrsp_hdr, which is currently named "padding"

If the RXR_PKT_CONNID_HDR flag is togged in "flags", this
space is used to store "connid" (sender's connection ID)

The patch then set "connid" and toggled the RXR_PKT_CONNID_HDR
flag in "flags".

It added a section to rxr_pkt_connid_ptr() to return pointer
of connid if RXR_PKT_CONNID_HDR has been set.

The macro RXR_READRSP_HDR_SIZE was removed for simplicity.

Signed-off-by: Wei Zhang <wzam@amazon.com>