prov/efa: add connid to EOR packet header
This patch added an multi-purpose field "padding/connid"
to the end of EOR packet.

The field can be used as "connid" field when the RXR_PKT_CONNID_HDR
flag is toggled in "flags".

If the flag is not toggled, the field is used as a padding space.

This patch uses it as "connid", (thus set the RXR_PKT_CONNID_HDF flag)

It added a section to rxr_pkt_connid_ptr() to return connid
pointer in EOR packet if RXR_PKT_CONNID_HDR is set.

Signed-off-by: Wei Zhang <wzam@amazon.com>