prov/rxm: Separate definitions of buffer, eager, and packet sizes
rxm_buffer_size and rxm_eager_limit are used throughout the code
with specific meanings.  In particular, the eager_limit can mean
the max size of eager message protocol or the size of bounce buffers
(relative to the user).  The latter is confusing, since buffer_size
has a similar meaning, but its value is not suitable.

The buffer_size is equal to eager_limit + the size of an rxm_pkt.

Update the buffer_size to equal the size of the allocated buffer.
Eager size will correspond with the size related to the eager
protocol.  And add a new packet_size variable to record the full
size of allocated packets.

Update the code to reference the variable that provides the
best meaning on how the value is used.  This will allow
future patches to allow these values to differ.  For example, the
eager_limit and buffer_size will not need to match.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>