prov/rxm: Allow specifying the eager size separate from buffer size
For rxm + tcp, the eager size and buffer sizes are able to differ.
Tcp allows posting the user buffer directly with tcp, avoiding the
send side copies.  On the receiving side, the buffer size is only
needed in the case of unexpected messages, which we can handle
separately.  (In follow on patch).

This change will allow us to reduce the memory footprint when
running over tcp by avoiding buffer allocations that aren't used
in the common cases.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>