prov/efa: move release of rdm_peer to efa_conn_release()
Currently, rxr_ep_close() call efa_rdm_peer_rest() to clear peer,
this is wrong because rdm_peer is part of efa_conn_release and should
be release in efa_conn_release().

This move the call to efa_rdm_peer_reset() to efa_conn_release()
from rxr_ep_close().

To do that, this commit converted rdm_peer->robuf from a pointer
to a real member, so that is not allocated from rxr_ep->robuf_pool.
thus does not depend on rxr_ep.

Signed-off-by: Wei Zhang <wzam@amazon.com>