core/mr_cache: Force flush 1 LRU entry when flush requested
Commit 5169390 changed the cache LRU flushing logic by moving
the flush check to the start of the loop.  The result is that
flushing is not performed as long as the cache is below its
threshhold.  This results in cleanup failing to run when the
cache is being destroyed.

The fi_mr_test will report an assertion that there are
outstanding registrations without the cleanup occurring.
Modify the LRU loop, so that we always flush at least 1
entry when flushing is requested and there are entries
to flush.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>