core/mr_cache: Fix flush_lru setting passed to flush
In ofi_mr_cache_search, if there are dead regions that need to
be flushed, we call cache_flush, but pass in true to flush_lru.
The result is that when cleaning up dead regions, we will always
flush an entry from the cache.

To avoid this, only flush the lru if the cache is full (when
searching to add new entries).

Signed-off-by: Sean Hefty <sean.hefty@intel.com>