fabtests: Set OOB socket to nonblocking in init_oob
We need to continue to make progress on the main ep even when
we're using the OOB socket.  Call ft_sock_setup() to initialize
the socket properly.  This fixes a hang in fi_unexpected_msg
test that shows up with rxm CM changes.  However, the hang could
happen with any test that uses OOB exchanges and blocks on recv().

Signed-off-by: Sean Hefty <sean.hefty@intel.com>