prov/tcp: Flush bsock and shutdown socket in ep_disable
tcpx_ep_disable is called in certain error cases, such as
problems during the CM exchange, if the underlying socket
has been disconnected (error on socket recv), or if the
user calls ep_shutdown.  When the ep is disabled, all
message queues are flushed and current operations are
canceled.  However, the ep is still attached with the
CQ and attempts can be made to progress it.

There's a possibility that data may have been queued
in the buffered socket prior to the ep being disabled.
Since the current receive operation was discarded, any
buffered data would be treated as a new message.  This
can lead to hitting an assert in tcpx_cq_progress that
the ep state is still connected.

To avoid hitting the assert, discard any buffered data.
Also call shutdown on the socket to disable further transfers
(in the case we reached ep_disable through one of the
error paths).

Signed-off-by: Sean Hefty <sean.hefty@intel.com>