prov/rxd: do not remove rxd addr from map and core AV on removal
When removing an address from the AV, the rxd
address/dg address should not be removed as well.
This is because the EP could still be communicating
with a peer for an outstanding transfer (or a retry).

If removed and reinserted, rxd will handle the duplicate
rxd address and map it accordingly.

This fixes a rare failure with the fi_av_xfer test which
is caused by a race condition where a peer tries to send
a retry or ACK after the address was removed from the AV.

Signed-off-by: aingerson <alexia.ingerson@intel.com>