prov/verbs: Additional checks for CQ's signaling_vector
We have recently discovered that some network devices do not create
one completion vector per core (it happens for example with virtual functions
when SRIOV is enabled, but there could be some other devices).

According to the man page of fi_cq_open, libfabric expects the value
for signaling_vector to be in the range of (0..max cpu - 1).
But CQ creation may still fail if the device doesn't create one
completion vector for every core.

If the comp_vector is incorrect, the Infiniband provider would return EINVAL.
Debugging such errors is however really hard (at least for Mellanox adapters)
because the error is thrown by the kernel module and there is absolutely no
debug message. The patch adds an error message if the value for the signal
vector is incorrect (lower than 0 or higher than the number of completion
vectors supported by the device). Hopefully that will help developers.

The patch also ensures that fi_cq_attr::signaling_vector is processed
only if FI_AFFINITY is set in fi_cq_attr::flags. If FI_AFFINITY
is not set, the comp_vector is set to 0, which would let the
kernel to select the best core for CQ interrupts.

Signed-off-by: Sylvain Didelot <sdidelot@ddn.com>