prov/shm: fix communication with peer that was restarted
If a peer was restarted, the server needs to remap the region to
update the smr pointer. Otherwise, the persistent peer will update the
local pointer and not the actual peer.
To recognize if the peer has been restarted or not, send the PID in the
send_name protocol (exchanges smr region on the first send). If the
PID is different than the one currently mapped in the pointer, unlink
the old region and remap to an updated pointer.
To fully fix this, we also need to track and update outstanding transfers
to and from the peer to either resend or complete in error. Adding TODO for this.

Signed-off-by: aingerson <alexia.ingerson@intel.com>