prov/efa: introducing extra feature flag RXR_REQ_FEATURE_ZERO_COPY_RECEIVE
This flag indicates an endpoint is in zero copy receive mode.

It is set in handshake packet's extra feature flags.

Upon receiving a handshake packet from peer, an endpoint
can tell whether the peer is in zero copy receive mode
using this flag.

If the peer is in zero copy receive mode, the endpoint
need to make sure the message prefix size remain constant
throughout the communication.

Signed-off-by: Wei Zhang <wzam@amazon.com>