prov/efa: change the way pkt_entry matching with rx_entry in zero copy
In two-side communications, each received packet entry needs to be
matched with an rx_entry.

Currently, a rx_entry was inserted to ep->rx_list when application posted
buffer. It is later remove from ep->rx_list when matched with a received
packet entry.

For zero copy receive, this method is inefficient. Because a packet
entry was also created together with the rx_entry, and its x_entry
field is pointing to the rx_entry.

It is also error-prone, because this method does not guarentee
the received pkt_entry is matched with the rx_entry it was created
with.

This patch make the following changes to the way packet entry is
matched with rx_entry in zero copy mode:

1. In rxr_msg_generic_recv(), rx_entry was not inserted to
   ep->rx_list.

2. In rxr_pkt_get_msgrtm_rx_entry(), if a pkt_entry is constructed
   using user supplied buffer, return pkt_entry->x_entry.

Signed-off-by: Wei Zhang <wzam@amazon.com>