prov/shm: fix av_lookup
av_lookup was not working properly because it was returning the ep->region name,
not the name in the AV. This causes an issue if the EP has not been mapped yet.
since av_insert can succeed without the peer being mapped yet, av_lookup should
do the same.
This exposed other issues with address name storage inconsistencies. The name being
returned in fi_info->dst_addr/src_addr include the smr_prefix which means they
are being inserted into the AV with those prefixes. However, they were being
stored in the ep without the prefix, since the shm region name does not include
the prefix. This is confusing and causes issues.
This patch moves the prefix elimination to the shm util code since the only reason
it is being removed is because shm regions cannot have '/'s in them. All references
to the EP name in the shm provider that do not refer to the physical shm EP name
have been changed to include the prefix for consistency.

Signed-off-by: aingerson <alexia.ingerson@intel.com>