prov/efa: fix a bug that rx_entry was not queued when encounter RNR
Currently, when a rx_entry encountered RNR for its control
packet, it is put to ep->rx_queued_entry_list (for resend
the control packet) only if its state is RXR_RX_RECV.

The original intention of checking state is to avoid same
rx_entry to be put in ep->rx_queued_entry_list more than
once.

However, it caused an issue: if rx_entry->state is RXR_RX_MATCHED,
rx_entry was not put to ep->rx_queued_entry_list, causing
the control packet not being resend.

In fact, when a rx_entry encountered RNR for a control packet,
it cannot already be in ep->rx_queued_entry_list. This is because
rx_entry only send 1 control packet at a time. If the send is
successful, it will be removed from ep->rx_queued_entry_list.

So this patch address the issue by always queue rx_entry if
it encountered RNR, and an assertion that the state cannot
be RXR_RX_QUEUED_CTRL.

It also merged several RX states into

     RXR_RX_QUEUED_CTRL
     RXR_RX_QUEUED_EOR
     RXR_RX_QUEUED_CTS_RNR

one state:

     RXR_RX_QUEUED_CTRL

Signed-off-by: Wei Zhang <wzam@amazon.com>