[v1.13.x] prov/efa: eliminate the flag RXR_PEER_BACKED_OFF
The flag RXR_PEER_BACKED_OFF is used to ensure that for each
call to progress engine, the backoff timeout and rnr_ts is updated
only once, in case multiple packets encountered RNR for one call
to progress engine.

However, the situation of multiple RNR packets encountered in
one call to progress engine will not happen because progress
engine will stop at the first error completion from device.

Therefore this flag is useless, so this patch removes it to
simplify the code and improve efficiency.

Signed-off-by: Wei Zhang <wzam@amazon.com>

cherry-picked from 60f6c14c6