[v1.13.x] prov/efa: refactor rxr_cq_queue_pkt()
This patch makes some changes to rxr_cq_queue_pkt() so it
is eariler to understand.

First, the section labelled as "queue_pkt" was moved from
the end to the beginning of the function. This changes makes
it clear that a packet is always queued.

Second, in various placed, the statement "goto queue_pkt"
was replaced with "return" because packet has been queued.

Finally, doxygen style comments have been added to
the function header. Comments in the function were
always updated to clear confusion

Signed-off-by: Wei Zhang <wzam@amazon.com>

cherry-picked from a5c3d9c0b