prov/efa: identify and ignore received packets from a removed peer
This patch introduced a function rxr_pkt_update_addr_by_hdr(),
which update pkt_entry->addr according to raw address or connid
in header.

If the packet is from a destroyed peer or a removed peer,
the function will set pkt_entry->addr to FI_ADDR_NOTAVAIL,
so that the packet will be ignored.

Signed-off-by: Wei Zhang <wzam@amazon.com>