[v1.13.x] prov/efa: only write RNR error completion for send operation
a provider should only write RNR error completion for a send
operation if application want to receive error completion for
RNR.

However, currently it is not the case for EFA.

Current, if application set to receive RNR error completion
entry, any RNR packet error will caused an RNR error completion
to be written.

This is wrong because not all packets are associate with
a send operiaton. For example, EOR and CTS packets are associated
with receive operations, and HANDSHAKE packet is not associated with
any user operation.

To addres this issue, and to make the code more clear,
this patch did a series of refactors:

    1. the function rxr_cq_handle_error() was splitted into two
       functions rxr_pkt_handle_send_error() and
       rxr_pkt_handle_recv_error().

    2. rxr_pkt_handle_send_error() handle RNR differently
       according to pkt_entry's x_entry type. It only write
       RNR error completion if x_entry is tx_entry.

    3. All other places where rxr_cq_handle_error() was called to handle
       a general error, the call was replaced with a call to efa_eq_write_error().

Signed-off-by: Wei Zhang <wzam@amazon.com>

cherry-picked from bf158f4e4