[v1.13.x] prov/efa: rename rxr_pkt_entry->type to rxr_pkt_entry->alloc_type
rxr_pkt_entry has a member type, which is used to indicate where
the pkt_entry's memory reside. It is sometimes confused with
the pkt_entry's type in the header, such as RXR_PKT_EAGER_RTM.

To avoid confusion, this patch rename it alloc_type. The following
changes are made accordingly:

1. enum rxr_pkt_entry_type was renamed rxr_pkt_entry_alloc_type.

2. The enum value RXR_PKT_POSTED was replaced with 4 new enum values:
        RXR_PKT_FROM_EFA_TX_PKT_POOL,
        RXR_PKT_FROM_EFA_RX_PKT_POOL,
        RXR_PKT_FROM_SHM_TX_PKT_POOL,
        RXR_PKT_FROM_SHM_RX_PKT_POOL.

3. A new argument "alloc_type" was added to rxr_pkt_entry_alloc(), now
   pkt_entry's alloc_type was specified when it is allocated

4. The argument "type" was remove from rxr_pkt_entry_copy().

5. rxr_pkt_entry_release_rx() was simplied to use alloc_type to determine
   whether it is for EFA device or for SHM.

Signed-off-by: Wei Zhang <wzam@amazon.com>

cherry-picked from 076057afb