fabtests: Do not immediately kill server process
Some fabtests are run in server/client mode. In non-strict
mode, if both server and client returning FI_ENODATA or
or FI_ENOSYS, the test is marked as 'Notrun' instead of
'Fail'.

Currently, when the client exits with non-zero code, the
server is immediately killed. However, doing such might
make the test that's supposed to be 'Notrun' to be
reported as 'Fail', because there's no guarantee
that the server has already exited. Killing the server
while it's running will change its return code.

Fix the issue by leaving the server more time (sleep 2s)
to exit.

Signed-off-by: Jie Zhang <zhngaj@amazon.com>