prov/efa: add connid to ATOMRSP packet header
This patch introduced a new usage for the 4-byte space
in rxr_atomrsp_hdr, which currently is named `padding`.

If the flag RXR_PKT_CONNID_HDR is toggled, this space
is used as "connid" (sender connection ID).

The patch use the space to store sender's connection ID,
and toggle the RXR_PKT_CONNID_HDR.

It added a section to rxr_pkt_connid_ptr() to return pointer
of connid in ATOMRSP packet if RXR_PKT_CONNID_HDR has been set.

Signed-off-by: Wei Zhang <wzam@amazon.com>