prov/efa: efa_av_remove() remove peer with resources
According to libfabric standard, fi_av_remove() should

"""
removes a set of addresses from an address vector. All resources
associated with the indicated addresses are released
"""

Currently, when fi_av_remove() is called, EFA will NOT remove
an address if there are resources associated with it.

Instead, the function will return -FI_EBUSY.

This behavior does not fit libfabric standard.

This patch addresses the issue by removing the address and releasing
all resources associated with it when fi_av_remove() is called.

It also add a function header commenting how on the fly RX/TX
packets for a removed address was handled.

Signed-off-by: Wei Zhang <wzam@amazon.com>