[v1.13.x] hmem/ze: Enable loading ZE library with dlopen()
Add configure option '--enable-ze-dlopen' which, when set to true, lets
the ZE library be loaded with dlopen() instead of becoming a runtime
dependency of libfabric.so. This allows libfabric to be built with
ZE support and still be able to run on systems that don't have the ZE
library installed.

Similar work has already been done with hmem/cuda and hmem/rocr.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>

Cherry-picked from 0e466a3b3