[v1.13.x] prov/efa: check for and handle empty util_av->ep_list in efa_av
Currently, efa_av assumes util_av->ep_list is not empty. The assumption
is not always right. This patch make efa_av to check whether
util_av->ep_list is empty and handles the case when it is empty.

util_av->ep_list is used in two place:

First, it is used in efa_conn_rdm_init(). When this fucntion is called,
util_av->ep_list cannot be empty, thus this patch added an assertion to
confirm that.

Second, it is used in efa_conn_rdm_deinit(). When this function is called,
util_av->ep_list can be emtpy. This patch adjust the code such that
ep was set to NULL when util_av->ep_list is empty, and make
efa_rdm_peer_clear() to be able to handle NULL ep.

Signed-off-by: Wei Zhang <wzam@amazon.com>

cherry-picked from 8f5efdc70