fabtests: Bug fixes related to IPv6 address format
The address format can be specified with the "-F" option, which is checked with
strncasecmp() in a predfined order. Previously the "fi_sockaddr_in" format was
checked before the "fi_sockaddr_in6" format, resulting the latter never being
reached. Fix by reversing the order of the two.

In addtion, the option was previously used by only 3 tests: "fi_msg_pingpong",
"fi_msg_bw" and "fi_rma_bw". Now expand the support to all the tests under the
"benchmark" directory.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>