prov/efa: EFA RDM sub-protocol name clean up
Currently, for message,emulated write and emulated read,
there is a protocol named "long". The name is to general
because there can be multiple ways to implement a long
message protocol. This patch rename them long-CTS protocol,
because what unique about this protocol is that this protocol
uses CTS packets to do flow control.

For, message and emulated write, there is a protocol named
"read". This patch renamed it to "long-read", to following the
naming convesion. "long" means it is intended for long messages,
"read" is how it is implemented (via RDMA read).

These changes conform to the document that describle EFA
RDM protocol v4.

The changes in this patch was generated using a script to
apply the following changes to all source file:

    s/_long_rtm/_longcts_rtm/g
    s/_long_msgrtm/_longcts_msgrtm/g
    s/_long_tagrtm/_longcts_tagrtm/g
    s/_LONG_MSGRTM/_LONGCTS_MSGRTM/g
    s/_LONG_TAGRTM/_LONGCTS_TAGRTM/g
    s/_read_rtm/_longread_rtm/g
    s/_read_msgrtm/_longread_msgrtm/g
    s/_read_tagrtm/_longread_tagrtm/g
    s/_READ_MSGRTM/_LONGREAD_MSGRTM/g
    s/_READ_TAGRTM/_LONGREAD_TAGRTM/g
    s/_long_rtw/_longcts_rtw/g
    s/_LONG_RTW/_LONGCTS_RTW/g
    s/_read_rtw/_longread_rtw/g
    s/_READ_RTW/_LONGREAD_RTW/g
    s/_long_rtr/_longcts_rtr/g
    s/_LONG_RTR/_LONGCTS_RTR/g

Signed-off-by: Wei Zhang <wzam@amazon.com>