prov/efa: use IBV_SEND_INLINE when message size fits
Currently, the IBV_SEND_INLINE flag was added when
FI_INJECT flag is used. This is inappropriate for
two reasons:

First, IBV_SEND_INLINE requires message size to
be less than or equal to maximum inline size, but
there is no check for that.

Second, IBV_SEND_INLINE is different from inject
because inject means no CQ entry, but using IBV_SEND_INLINE
still generates an CQ entry.

This patch change it to that the IBV_SEND_INLINE flag
is used when the message size is less than or equal to
inline buffer size in ibv_ctx.

Signed-off-by: Wei Zhang <wzam@amazon.com>